Rename compilability test classes to ThisShouldCompile and delete tests
