+++ test-dependent-projects/java-dep-webauthn-server-attestation/build.gradle.kts
@@ -4,7 +4,5 @@ plugins {
 
 dependencies {
     implementation(project(":webauthn-server-attestation"))
-
-    testImplementation("junit:junit:4.12")
 }
