--- test-dependent-projects/java-dep-webauthn-server-attestation/src/test/java/com/yubico/example/ExampleTest.java
@@ -1,14 +1,0 @@
-package com.yubico.example;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertNotNull;
-
-public class ExampleTest {
-
-    @Test
-    public void resolverIsNotNull() {
-        assertNotNull(new Example().getResolver());
-    }
-
-}
