+++ test-dependent-projects/java-dep-webauthn-server-core/build.gradle.kts
@@ -4,7 +4,5 @@ plugins {
 
 dependencies {
     implementation(project(":webauthn-server-core"))
-
-    testImplementation("junit:junit:4.12")
 }
