+++ test-dependent-projects/java-dep-yubico-util/src/main/java/com/yubico/test/compilability/ThisShouldCompile.java
@@ -1,9 +1,9 @@
-package com.yubico.example;
+package com.yubico.test.compilability;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.yubico.internal.util.JacksonCodecs;
 
-public class Example {
+public class ThisShouldCompile {
 
     public String getEncodedValue() throws JsonProcessingException {
         return JacksonCodecs.json().writeValueAsString("hej");
